% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dayoff-package.R
\docType{package}
\name{dayoff-package}
\alias{dayoff}
\alias{dayoff-package}
\title{dayoff: Bioinformatics in R}
\description{
Bioinformatics in R.
}
\author{
\strong{Maintainer}: Nathan Brouwer \email{brouwern@gmail.com}

}
\keyword{internal}
