% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta_cleaner.R
\name{fasta_cleaner}
\alias{fasta_cleaner}
\title{Function to convert a FASTA file stored as an object into a vector}
\usage{
fasta_cleaner(fasta_object, parse = TRUE)
}
\description{
Function to convert a FASTA file stored as an object into a vector
}
\examples{

header. <- ">NC_001477.1 Dengue virus 1, complete genome"
fasta_cleaner(dengueseq_fasta, parse = TRUE)

}
