% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hShroom2.R
\docType{data}
\name{hShroom2}
\alias{hShroom2}
\title{Human Shoom 2 Fasta File}
\format{An object of class \code{character} of length 1.}
\usage{
hShroom2
}
\description{
Human Shoom 2 Fasta File
}
\keyword{datasets}
