% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hShroom3.R
\docType{data}
\name{hShroom3}
\alias{hShroom3}
\title{Human Shroom 3 Fasta File}
\format{An object of class \code{character} of length 1.}
\usage{
hShroom3
}
\description{
Human Shroom 3 Fasta File
}
\keyword{datasets}
