% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mShroom3a.R
\docType{data}
\name{mShroom3a}
\alias{mShroom3a}
\title{Mouse Shroom 3a Fasta File}
\format{An object of class \code{character} of length 1.}
\usage{
mShroom3a
}
\description{
Mouse Shroom 3a Fasta File
}
\keyword{datasets}
