% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shroom_table.R
\docType{data}
\name{shroom_table}
\alias{shroom_table}
\title{Table of shroom accessions numbers circa 2006}
\format{An object of class \code{data.frame} with 14 rows and 4 columns.}
\usage{
shroom_table
}
\description{
Table of shroom accessions numbers circa 2006
}
\keyword{datasets}
