% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sShroom.R
\docType{data}
\name{sShroom}
\alias{sShroom}
\title{Sea Urchin Shoom Fasta File}
\format{An object of class \code{character} of length 1.}
\usage{
sShroom
}
\description{
Sea Urchin Shoom Fasta File
}
\keyword{datasets}
