% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutor_open.R
\name{tutor_open}
\alias{tutor_open}
\title{Open a specified tutorial file}
\usage{
tutor_open(tutorial = NULL)
}
\arguments{
\item{tutorial}{name of tutorial to open}
}
\description{
Open a specified tutorial file
}
\examples{
tutor_open()

}
